requirements.txt
Flask-OIDC==2.2.0
Flask-OpenID==1.3.0
mysqlclient==2.2.4
psycopg2-binary==2.9.9
werkzeug==2.3.7
python-openid==2.2.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-login==0.6.3
flask-wtf==1.2.1
flask-babel==2.0.0
flask-migrate==4.0.5
flask-sqlalchemy==2.5.1
Pillow==10.2.0
redis==5.0.1
flask-limiter==3.5.0 